requirements.txt
Django==4.2
django-cors-headers==3.5.0
django-dotenv==1.4.2
djangorestframework==3.14.0
torch==1.9.1
transformers==4.12.5
datasets==1.16.1
tokenizers==0.10.3